etf_beta-5

+++ strategies/etf_beta.py
@@ -13,7 +13,7 @@ def trade(exchange):
         return []
     ave = 3 * np.mean(bond[-5:]) + 2 * np.mean(gs[:-5]) + 3 * np.mean(ms[:-5]) + 2 * np.mean(wfc[:-5])
     if (xlf < ave):
-        trades.append(("BUY", "XLF", 1.3*ave, 30))
+        trades.append(("BUY", "XLF", (int)(1.3*ave), 30))
     else:
-        trades.append(("SELL", "XLF", 0.8*xlf, 30))
+        trades.append(("SELL", "XLF", (int)(0.8*xlf), 30))
     return trades
